backend/requirements.txt
# Core Framework
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
alembic

# AWS SDK
boto3
botocore

# Database drivers
pymysql  # MySQL (optional)
# psycopg2-binary  # PostgreSQL (uncomment if needed, requires pg_config on Windows)

# Authentication
python-jose
passlib
python-multipart

# HTTP client
httpx
requests

# Google Gemini AI
google-generativeai>=0.3.0

# Social Media SDKs
tweepy>=4.14.0  # Twitter / X API
